issue: 4207451 Ensure type safety for config vals
Replace all instances of registry.get_value<int64_t> with
registry.get_value<decltype(var)> throughout the codebase.

This change:
- Improves type safety - ensure config values match target variable type
- Eliminates manual range checking and explicit casting code
- Reduces risk of data corruption from improper type conversion
- Makes the code more concise and easier to maintain

This change simplifies many lines of code while improving safety, as any
type mismatch will now be caught at compile time rather than requiring
manual runtime validation.

As part of a representation limitation of config provider, I've added
a new unit test that will fail if a size_t value is requested as it's
not supported, as integers are represented as int64_t in the config.

+++ src/core/dev/hw_queue_tx.cpp
@@ -1019,7 +1019,7 @@ std::unique_ptr<dpcp::tls_dek> hw_queue_tx::get_tls_dek(const void *key, uint32_
     // of active connections.
     if (unlikely(!m_p_ring->tls_sync_dek_supported()) ||
         (unlikely(m_tls_dek_get_cache.empty()) &&
-         (m_tls_dek_put_cache.size() <= safe_mce_sys().utls_low_wmark_dek_cache_size))) {
+         (static_cast<int64_t>(m_tls_dek_put_cache.size()) <= safe_mce_sys().utls_low_wmark_dek_cache_size))) {
         return get_new_tls_dek(key, key_size_bytes);
     }
 
@@ -1061,7 +1061,7 @@ void hw_queue_tx::put_tls_dek(std::unique_ptr<dpcp::tls_dek> &&tls_dek_obj)
     }
     // We don't allow unlimited DEK cache to avoid system DEK starvation.
     if (likely(m_p_ring->tls_sync_dek_supported()) &&
-        m_tls_dek_put_cache.size() < safe_mce_sys().utls_high_wmark_dek_cache_size) {
+        static_cast<int64_t>(m_tls_dek_put_cache.size()) < safe_mce_sys().utls_high_wmark_dek_cache_size) {
         m_tls_dek_put_cache.emplace_back(std::forward<std::unique_ptr<dpcp::tls_dek>>(tls_dek_obj));
     }
 }
